bookmark backend 대규모 개편 및 cache 기능 추가

+++ src/main/java/com/pjh/bookmark/service/BookmarkService.java
@@ -1,11 +1,13 @@
 package com.pjh.bookmark.service;
 
-import com.pjh.bookmark.dto.BookmarkRequestDto;
 import com.pjh.bookmark.entity.Bookmark;
+import com.pjh.bookmark.exception.BookmarkException;
 import com.pjh.bookmark.exception.SuccessException;
 import com.pjh.bookmark.exception.UnExpectedException;
 import com.pjh.bookmark.repository.BookmarkRepository;
 import com.pjh.bookmark.repository.HashMapRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +20,7 @@ public class BookmarkService {
     private static final int LIVE_BOOKMARK_STATE = 1;
     private static final int DEAD_BOOKMARK_STATE = 0;
     private static final int MAIN_BOOKMARK_NUM = 1;
+    private final Logger logger = LoggerFactory.getLogger(this.getClass().getSimpleName());
 
     @Autowired
     private BookmarkRepository bookmarkRepository;
@@ -37,36 +40,39 @@ public class BookmarkService {
     }
 
     public void addBookmarkFrequencyFunc(long bookmarkId, long userId){
-        Bookmark bookmark = bookmarkRepository.findByBookmarkIdAndUserId(bookmarkId, userId);
+        Bookmark bookmark = bookmarkRepository.findByBookmarkIdAndUserIdAndState(bookmarkId, userId, LIVE_BOOKMARK_STATE);
         if( bookmark != null){
             bookmark.setFrequency(bookmark.getFrequency() + 1);
             bookmarkRepository.save(bookmark);
             throw new SuccessException("freq add success");
         }
         else {
-            throw new UnExpectedException("freq add fail");
+            throw new BookmarkException("freq add fail");
         }
     }
 
-    public Bookmark createBookmarkFunc(BookmarkRequestDto bookmarkRequestDto, long userId){
-        Bookmark saveBookmark = bookmarkRequestDto.getBookmark();
+    public Bookmark createBookmarkFunc(Bookmark bookmark, long userId){
+        Bookmark saveBookmark = bookmark;
         saveBookmark.setUserId(userId);
+        saveBookmark.setFrequency(0);
         return bookmarkRepository.save(saveBookmark);
     }
 
-    public Bookmark updateBookmarkFunc(BookmarkRequestDto bookmarkRequestDto, long userId){
-        //TODO: 하나만 업데이트 할 수 있도록 수정 필요
-        Bookmark updateBookmark = bookmarkRequestDto.getBookmark();
-        updateBookmark.setUserId(userId);
-        updateBookmark.setState(LIVE_BOOKMARK_STATE);
+    public Bookmark updateBookmarkFunc(Bookmark bookmark, long userId){
+        Bookmark updateBookmark = bookmarkRepository.findByBookmarkIdAndUserIdAndState(bookmark.getBookmarkId(), userId, LIVE_BOOKMARK_STATE);
+        if (updateBookmark == null) throw new BookmarkException("update bookmark not found");
+        if (bookmark.getBookmarkIcon() != null && !bookmark.getBookmarkIcon().equals("")) updateBookmark.setBookmarkIcon(bookmark.getBookmarkIcon());
+        if (bookmark.getBookmarkInfo() != null && !bookmark.getBookmarkInfo().equals("")) updateBookmark.setBookmarkInfo(bookmark.getBookmarkInfo());
+        if (bookmark.getBookmarkName() != null && !bookmark.getBookmarkName().equals("")) updateBookmark.setBookmarkName(bookmark.getBookmarkName());
+        if (bookmark.getUrl() != null && !bookmark.getUrl().equals("")) updateBookmark.setUrl(bookmark.getUrl());
         return bookmarkRepository.save(updateBookmark);
     }
 
     public Bookmark deleteBookmarkFunc(long bookmarkId, long userId){
-        Bookmark deleteBookmark = bookmarkRepository.findByBookmarkIdAndUserId(bookmarkId, userId);
+        Bookmark deleteBookmark = bookmarkRepository.findByBookmarkIdAndUserIdAndState(bookmarkId, userId, LIVE_BOOKMARK_STATE);
+        if (deleteBookmark == null) throw new BookmarkException("delete bookmark not found");
         hashService.deleteHashMapByBookmarkFunc(deleteBookmark.getBookmarkId());
         deleteBookmark.setState(DEAD_BOOKMARK_STATE);
-        deleteBookmark.setUserId(userId);
         return bookmarkRepository.save(deleteBookmark);
     }
 
@@ -77,7 +83,7 @@ public class BookmarkService {
 
         List<Bookmark> bookmarkList = new ArrayList<>();
         for (Long bookmarkId : bookmarkIdListInBookmark ) {
-            Bookmark bookmark = bookmarkRepository.findByBookmarkIdAndUserId(bookmarkId, userId);
+            Bookmark bookmark = bookmarkRepository.findByBookmarkIdAndUserIdAndState(bookmarkId, userId, LIVE_BOOKMARK_STATE);
             if ( bookmark != null ) {
                 bookmarkList.add(bookmark);
             }
